backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.3
RestrictedPython==7.0

# For data processing
pandas==2.1.4
numpy==1.26.2

# For NLP
nltk==3.8.1
